+++ Team9853/src/main/java/org/firstinspires/ftc/team9853/opmodes/teleop/TeleTest.java
@@ -1,8 +1,9 @@
 package org.firstinspires.ftc.team9853.opmodes.teleop;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-/**
+/*!
  * FTC_APP_2018
  * Copyright (c) 2017 Chatham Robotics
  * MIT License
@@ -10,5 +11,16 @@ import com.qualcomm.robotcore.eventloop.opmode.OpMode;
  * @Last Modified by: storm
  * @Last Modified time: 9/17/2017
  */
+@TeleOp(name = "Tele Test")
+@SuppressWarnings("unused")
 public class TeleTest extends OpMode {
+    @Override
+    public void init() {
+
+    }
+
+    @Override
+    public void loop() {
+
+    }
 }
